requirements.txt
coverage==7.0.1
flask==2.2.2
flask_migrate==4.0.0
flask_sqlalchemy==3.0.2
pandas==1.5.2
parameterized==0.8.1
psycopg2==2.9.5
pytest==7.2.0
python-dotenv==0.21.0
tomli==2.0.1